requirements.txt
fastapi
uvicorn
requests
python-dotenv
yt-dlp
pydantic
scenedetect[opencv]
ultralytics
opencv-python
numpy
